lib/Cell_DETR_master/requirements.txt
pytorch==1.4.0
torchvision==0.5.0
numpy
skimage
scipy
tqdm
setproctitle
tikzplotlib